Edit部分を別ファイルに切り出し files stateのリフトアップ
